api/requirements.txt
azure-devops==6.0.0b4
azure-functions==1.8.0
certifi==2021.10.8
charset-normalizer==2.0.11
idna==3.3
isodate==0.6.1
msrest==0.6.21
oauthlib==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
urllib3==1.26.8
